work page also moved to seperate UI
