+++ lib/utilities/open_functions.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:portfolio/l10n/l10n.dart';
 import 'package:portfolio/utilities/app_cinfiguration.dart';
+import 'package:portfolio/work/views/components/additional_projects.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 void appsLaunchGivenUrl({required String urlToLaunch}) {
@@ -74,4 +75,44 @@ FaIcon _socialIcon({required IconData icon}) {
     icon,
     color: Colors.white,
   );
+}
+
+List<Widget> additionalProjects(
+  BuildContext context,
+  AppLocalizations l10n,
+) {
+  return <Widget>[
+    AdditionalProject(
+      context: context,
+      projectIcon: Icons.voice_chat_outlined,
+      projectUrl: l10n.workVoiceRecorderUrl,
+      projectName: l10n.workVoiceRecorderTitle,
+      projectDetail: l10n.workVoiceRecorderDetail,
+      technologies: l10n.workVoiceRecorderTechnologies,
+    ),
+    AdditionalProject(
+      context: context,
+      projectIcon: Icons.folder_outlined,
+      projectUrl: l10n.workAuthBlocUrl,
+      projectName: l10n.workAuthBlocTitle,
+      projectDetail: l10n.workAuthBlocDetail,
+      technologies: l10n.workAuthBlocTechnologies,
+    ),
+    AdditionalProject(
+      context: context,
+      projectIcon: Icons.image_outlined,
+      projectUrl: l10n.workGifMakerUrl,
+      projectName: l10n.workGifMakerTitle,
+      projectDetail: l10n.workGifMakerDetail,
+      technologies: l10n.workGifMakerTechnologies,
+    ),
+    AdditionalProject(
+      context: context,
+      projectIcon: Icons.chat_outlined,
+      projectUrl: l10n.workFlutterChatUrl,
+      projectName: l10n.workFlutterChatTitle,
+      projectDetail: l10n.workFlutterChatDetail,
+      technologies: l10n.workFlutterChatTechnologies,
+    )
+  ];
 }
